+++ client/src/App.js
@@ -4,7 +4,7 @@ import { useDispatch } from 'react-redux'
 import Box from '@material-ui/core/Box'
 import { makeStyles } from '@material-ui/core/styles'
 
-import Header from './components/HeaderNew'
+import Header from './components/Header'
 import MainRoutes from './routes/MainRoutes'
 import Footer from './components/Footer'
 import Modal from './components/Modal'
